feat: Add ImageAliasNotFound exception and update image repository methods for better error handling
