+++ src/ai/backend/manager/services/image/actions/alias_image.py
@@ -2,15 +2,6 @@ from dataclasses import dataclass
 from typing import Optional, override
 from uuid import UUID
 
-from aiohttp import web
-
-from ai.backend.common.exception import (
-    BackendAIError,
-    ErrorCode,
-    ErrorDetail,
-    ErrorDomain,
-    ErrorOperation,
-)
 from ai.backend.manager.actions.action import BaseActionResult
 from ai.backend.manager.data.image.types import ImageAliasData
 from ai.backend.manager.services.image.actions.base import ImageAction
@@ -40,33 +31,3 @@ class AliasImageActionResult(BaseActionResult):
     @override
     def entity_id(self) -> Optional[str]:
         return str(self.image_id)
-
-
-class AliasImageActionValueError(BackendAIError, web.HTTPBadRequest):
-    error_type = "https://api.backend.ai/probs/invalid-parameters"
-    error_title = "Invalid parameters for image alias."
-
-    @classmethod
-    def error_code(cls) -> ErrorCode:
-        return ErrorCode(
-            domain=ErrorDomain.IMAGE_ALIAS,
-            operation=ErrorOperation.CREATE,
-            error_detail=ErrorDetail.INVALID_PARAMETERS,
-        )
-
-
-class AliasImageActionDBError(BackendAIError, web.HTTPInternalServerError):
-    """
-    This can occur when an image alias with the same value already exists.
-    """
-
-    error_type = "https://api.backend.ai/probs/image-db-error"
-    error_title = "Database error while managing image alias."
-
-    @classmethod
-    def error_code(cls) -> ErrorCode:
-        return ErrorCode(
-            domain=ErrorDomain.IMAGE_ALIAS,
-            operation=ErrorOperation.UPDATE,
-            error_detail=ErrorDetail.ALREADY_EXISTS,
-        )
