+++ src/ai/backend/manager/services/image/actions/dealias_image.py
@@ -2,15 +2,6 @@ from dataclasses import dataclass
 from typing import Optional, override
 from uuid import UUID
 
-from aiohttp import web
-
-from ai.backend.common.exception import (
-    BackendAIError,
-    ErrorCode,
-    ErrorDetail,
-    ErrorDomain,
-    ErrorOperation,
-)
 from ai.backend.manager.actions.action import BaseActionResult
 from ai.backend.manager.data.image.types import ImageAliasData
 from ai.backend.manager.services.image.actions.base import ImageAction
@@ -38,16 +29,3 @@ class DealiasImageActionResult(BaseActionResult):
     @override
     def entity_id(self) -> Optional[str]:
         return str(self.image_id)
-
-
-class DealiasImageActionNoSuchAliasError(BackendAIError, web.HTTPNotFound):
-    error_type = "https://api.backend.ai/probs/image-alias-not-found"
-    error_title = "Image alias not found."
-
-    @classmethod
-    def error_code(cls) -> ErrorCode:
-        return ErrorCode(
-            domain=ErrorDomain.IMAGE_ALIAS,
-            operation=ErrorOperation.UPDATE,
-            error_detail=ErrorDetail.NOT_FOUND,
-        )
