+++ src/ai/backend/manager/services/image/actions/forget_image_by_id.py
@@ -2,13 +2,6 @@ import uuid
 from dataclasses import dataclass
 from typing import Optional, override
 
-from ai.backend.common.exception import (
-    BackendAIError,
-    ErrorCode,
-    ErrorDetail,
-    ErrorDomain,
-    ErrorOperation,
-)
 from ai.backend.manager.actions.action import BaseActionResult
 from ai.backend.manager.data.image.types import ImageData
 from ai.backend.manager.models.user import UserRole
@@ -38,16 +31,3 @@ class ForgetImageByIdActionResult(BaseActionResult):
     @override
     def entity_id(self) -> Optional[str]:
         return str(self.image.id)
-
-
-class ForgetImageActionByIdGenericForbiddenError(BackendAIError):
-    error_type = "https://api.backend.ai/probs/generic-forbidden"
-    error_title = "Access to this resource is forbidden."
-
-    @classmethod
-    def error_code(cls) -> ErrorCode:
-        return ErrorCode(
-            domain=ErrorDomain.IMAGE,
-            operation=ErrorOperation.SOFT_DELETE,
-            error_detail=ErrorDetail.FORBIDDEN,
-        )
